Create some functionality of page 3
Added a new event listener that removes content paragraphs
and most of the footer. Set the input box similar to the
name box but it will react only when supplied integer.

Couldn't figure out why the console throws errors when I
press Enter. For some reason, the handler function doesn't
replace <kbd> with a text node.